// app-admin/src/types/database.types.ts
export interface Etablissement {
  id: string;
  nom: string;
  adresse: string | null;
  telephone: string | null;
  email: string | null;
  statut_abonnement: 'actif' | 'expire' | 'suspendu';
  date_debut: string;
  date_fin: string;
  actif: boolean;
  dernier_paiement_date: string | null;
  dernier_paiement_confirme_par: string | null;
  date_creation: string;
  date_modification: string;
}